actual refactor
